refactor(fbuild-config): split board.rs into submodules to satisfy LOC gate
